+++ src/components/Project/Project.tsx
@@ -1,8 +1,20 @@
+import { JSX } from 'react'
 import GDG from './Gdg'
 import Jinlo from './Jinlo'
 import Ladi from './Ladi'
 import TodayClothes from './TodayClothes'
 import TodayScrum from './TodayScrum'
+
+export type CardData = {
+  id: string
+  title: string
+  color: string
+  description?: string
+  imageUrl?: string
+  tags?: string[]
+  component: JSX.Element
+  team: boolean
+}
 
 const PROJECTS = [
   {
@@ -12,6 +24,7 @@ const PROJECTS = [
     description: 'AI 기반 코드 리뷰 서비스',
     tags: ['앱', 'AI'],
     component: <Jinlo />,
+    team: true,
   },
   {
     color: 'bg-[#9097C0]',
@@ -21,6 +34,7 @@ const PROJECTS = [
     imageUrl: '/images/project/Gdg/logo.gif',
     tags: ['웹/앱', '플리마켓'],
     component: <GDG />,
+    team: true,
   },
   {
     color: 'bg-[#1666DB]',
@@ -30,6 +44,7 @@ const PROJECTS = [
     imageUrl: '/images/project/ladi/logo.png',
     tags: ['웹/앱', '반응형'],
     component: <Ladi />,
+    team: false,
   },
   {
     color: 'bg-[#f47272]',
@@ -38,6 +53,7 @@ const PROJECTS = [
     description: 'AI를 활용한 오늘의 옷 추천 서비스',
     tags: ['AI', '웹/앱', '반응형'],
     component: <TodayClothes />,
+    team: true,
   },
   {
     color: 'bg-[#d167f5]',
@@ -46,6 +62,7 @@ const PROJECTS = [
     description: '프로젝트 관리 및 팀원간 소통을 위한 스크럼 서비스',
     tags: ['웹'],
     component: <TodayScrum />,
+    team: true,
   },
 ]
